Changes to file structure to follow a common structure with zephyr component.
