+++ esp_idf/spotflow/CMakeLists.txt
@@ -1,4 +1,4 @@
-idf_component_register(SRCS "spotflow.c" "spotflow_mqtt.c"
+idf_component_register(SRCS "srcs/net/spotflow_mqtt.c" "srcs/logging/spotflow_log_backend.c" "srcs/logging/spotflow_log_cbor.c" "srcs/logging/spotflow_log_json.c" "spotflow.c"
                     PRIV_REQUIRES mqtt esp_partition nvs_flash esp_netif app_update
                     INCLUDE_DIRS "include"
                     REQUIRES cbor
