+++ esp_idf/spotflow/include/logging/spotflow_log_backend.h
@@ -1,6 +1,9 @@
 #ifndef SPOTFLOW_LOG_H
 #define SPOTFLOW_LOG_H
 
+#include "spotflow.h"
+
+int spotflow_log_backend(const char *fmt, va_list args);
 #if CONFIG_SPOTFLOW_DEBUG_MESSAGE_TERMINAL
     #define SPOTFLOW_LOG(fmt, ...) printf("[SPOTFLOW] " fmt, ##__VA_ARGS__)
 #else
