+++ esp_idf/spotflow/include/logging/spotflow_log_json.h
@@ -1,0 +1,8 @@
+#ifndef SPOTFLOW_LOG_JSON_H
+#define SPOTFLOW_LOG_JSON_H
+
+#include "spotflow.h"
+
+char* log_json(char* body, const char* severity);
+
+#endif
