+++ esp_idf/spotflow/include/spotflow.h
@@ -1,15 +1,34 @@
 #ifndef SPOTFLOW_H
 #define SPOTFLOW_H
 
+#include <stdio.h>
+#include <stdint.h>
+#include <stddef.h>
+#include <string.h>
+#include "esp_system.h"
+#include "esp_partition.h"
+#include "nvs_flash.h"
+#include "esp_event.h"
+#include "esp_netif.h"
 
-#include "spotflow_mqtt.h"
-#include "spotflow_log.h"
+#include "esp_log.h"
+#include "esp_mac.h"
+#include "mqtt_client.h"
+#include "esp_tls.h"
+#include <sys/param.h>
+#include <time.h>
+#include "esp_log_timestamp.h"
 
-#if CONFIG_SPOTFLOW_DEBUG
-    #define SPOTFLOW_PRINTF(fmt, ...)   printf("[SPOTFLOW] " fmt, ##__VA_ARGS__)
+#include "net/spotflow_mqtt.h"
+#include "logging/spotflow_log_backend.h"
+
+#ifndef CONFIG_USE_JSON_PAYLOAD
+    #include "logging/spotflow_log_cbor.h"
 #else
-    #define SPOTFLOW_PRINTF(fmt, ...)   do {} while(0)
+    #include "logging/spotflow_log_json.h"
 #endif
+
+extern vprintf_like_t original_vprintf;
 void spotflow_init(void);
 
 #endif
